+++ packages/xchain-mayachain-query/__tests__/mayachain-query.test.ts
@@ -14,7 +14,7 @@ import {
 
 import mockMayanodeApi from '../__mocks__/mayanode-api'
 import mockMayamidgardApi from '../__mocks__/midgard-api'
-import { BtcAsset, EthAsset, RuneAsset } from '../src/'
+import { BtcAsset, EthAsset, PendingSwap, RuneAsset, SuccessSwap } from '../src/'
 import { MayachainCache } from '../src/mayachain-cache'
 import { MayachainQuery } from '../src/mayachain-query'
 
@@ -181,37 +181,101 @@ describe('Mayachain-query tests', () => {
   it('Should get swaps history', async () => {
     const swapResume = await mayachainQuery.getSwapHistory({ addresses: ['address'] })
     expect(swapResume.count === swapResume.swaps.length)
+    expect(swapResume.count).toBe(4)
+    const pendingSwap = swapResume.swaps[0] as PendingSwap
     expect({
-      date: swapResume.swaps[0].date,
-      status: swapResume.swaps[0].status,
+      date: pendingSwap.date,
+      status: pendingSwap.status,
+      fromAsset: assetToString(pendingSwap.fromAsset),
+      toAsset: assetToString(pendingSwap.toAsset),
       in: {
-        hash: swapResume.swaps[0].inboundTx.hash,
-        address: swapResume.swaps[0].inboundTx.address,
-        asset: assetToString(swapResume.swaps[0].inboundTx.amount.asset),
-        amount: baseToAsset(swapResume.swaps[0].inboundTx.amount.baseAmount).amount().toString(),
+        hash: pendingSwap.inboundTx.hash,
+        address: pendingSwap.inboundTx.address,
+        asset: assetToString(pendingSwap.inboundTx.amount.asset),
+        amount: baseToAsset(pendingSwap.inboundTx.amount.baseAmount).amount().toString(),
+      },
+    }).toEqual({
+      date: new Date('2024-09-20T11:46:24.224Z'),
+      status: 'pending',
+      fromAsset: 'ETH.USDT-0XDAC17F958D2EE523A2206206994597C13D831EC7',
+      toAsset: 'THOR.RUNE',
+      in: {
+        hash: '4C8A83C316B0247AE293F3D5A0B07974DE4083F220BF5551ACF20E66D17BC252',
+        address: '0xd1f7112354055160d58fa1b1e7cfd15c0bfee464',
+        asset: 'ETH.USDT-0XDAC17F958D2EE523A2206206994597C13D831EC7',
+        amount: '104.34782608',
+      },
+    })
+    let successSwap = swapResume.swaps[1] as SuccessSwap
+    expect({
+      date: successSwap.date,
+      status: successSwap.status,
+      fromAsset: assetToString(successSwap.fromAsset),
+      toAsset: assetToString(successSwap.toAsset),
+      in: {
+        hash: successSwap.inboundTx.hash,
+        address: successSwap.inboundTx.address,
+        asset: assetToString(successSwap.inboundTx.amount.asset),
+        amount: baseToAsset(successSwap.inboundTx.amount.baseAmount).amount().toString(),
       },
       out: {
-        hash: swapResume.swaps[0].outboundTx?.hash,
-        address: swapResume.swaps[0].outboundTx?.address,
-        asset: swapResume.swaps[0].outboundTx ? assetToString(swapResume.swaps[0].outboundTx.amount.asset) : undefined,
-        amount: swapResume.swaps[0].outboundTx
-          ? baseToAsset(swapResume.swaps[0].outboundTx.amount.baseAmount).amount().toString()
-          : undefined,
+        hash: successSwap.outboundTx.hash,
+        address: successSwap.outboundTx.address,
+        asset: assetToString(successSwap.outboundTx.amount.asset),
+        amount: baseToAsset(successSwap.outboundTx.amount.baseAmount).amount().toString(),
       },
     }).toEqual({
       date: new Date('2024-03-12T02:28:28.760Z'),
       status: 'success',
+      fromAsset: 'ETH.ETH',
+      toAsset: 'MAYA.CACAO',
       in: {
         hash: '224CAF4D502A0A415F1312AFD16C0E7A2E3E79840AF593C2F875C806AA12E020',
         address: '0xaa278b62225f6dbc4436de8fa3dd195e1542d159',
         asset: 'ETH.ETH',
-        amount: '0.99',
+        amount: '1',
       },
       out: {
         hash: '',
         address: 'maya17xu9ej4rkxsmnl3wkp0kph6k4jk70gzay56p0l',
         asset: 'MAYA.CACAO',
         amount: '3329.7336036086',
+      },
+    })
+    successSwap = swapResume.swaps[2] as SuccessSwap
+    expect({
+      date: successSwap.date,
+      status: successSwap.status,
+      fromAsset: assetToString(successSwap.fromAsset),
+      toAsset: assetToString(successSwap.toAsset),
+      in: {
+        hash: successSwap.inboundTx.hash,
+        address: successSwap.inboundTx.address,
+        asset: assetToString(successSwap.inboundTx.amount.asset),
+        amount: baseToAsset(successSwap.inboundTx.amount.baseAmount).amount().toString(),
+      },
+      out: {
+        hash: successSwap.outboundTx.hash,
+        address: successSwap.outboundTx.address,
+        asset: assetToString(successSwap.outboundTx.amount.asset),
+        amount: baseToAsset(successSwap.outboundTx.amount.baseAmount).amount().toString(),
+      },
+    }).toEqual({
+      date: new Date('2024-05-08T10:14:19.088Z'),
+      status: 'success',
+      fromAsset: 'ETH.ETH',
+      toAsset: 'ARB.TGT-0X429FED88F10285E61B12BDF00848315FBDFCC341',
+      in: {
+        hash: '4CC77E1FC51E9F99F0EE2AAC42FA2FD38A149796E9AD0B8368A1B3EB80AA4C85',
+        address: '0xaa278b62225f6dbc4436de8fa3dd195e1542d159',
+        asset: 'ETH.ETH',
+        amount: '0.99',
+      },
+      out: {
+        hash: '1E9F450E6174407FE846929BD16D057547DFB3BD1918176E12AACDAF6C753A8B',
+        address: '0xAA278b62225f6dbc4436De8Fa3dD195e1542d159',
+        asset: 'ARB.TGT-0X429FED88F10285E61B12BDF00848315FBDFCC341',
+        amount: '77208.44802738',
       },
     })
   })