+++ packages/xchain-mayachain-query/src/mayachain-query.ts
@@ -1,10 +1,13 @@
 import { Network } from '@xchainjs/xchain-client'
-import { PoolDetail, Transaction } from '@xchainjs/xchain-mayamidgard'
+import { Action, PoolDetail, SwapMetadata, Transaction } from '@xchainjs/xchain-mayamidgard'
 import { QuoteSwapResponse } from '@xchainjs/xchain-mayanode'
 import {
   Address,
   AssetCryptoAmount,
   CryptoAmount,
+  SYNTH_ASSET_DELIMITER,
+  TOKEN_ASSET_DELIMITER,
+  TRADE_ASSET_DELIMITER,
   assetFromStringEx,
   assetToString,
   baseAmount,
@@ -23,6 +26,7 @@ import {
   QuoteSwapParams,
   SwapHistoryParams,
   SwapsHistory,
+  TransactionAction,
 } from './types'
 import {
   ArbAsset,
@@ -285,6 +289,7 @@ export class MayachainQuery {
       type: 'swap',
     })
     const assetDecimals = await this.mayachainCache.getAssetDecimals()
+    const poolDetails = await this.mayachainCache.getPools()
 
     const getCryptoAmount = (
       assetDecimals: Record<string, number>,
@@ -300,32 +305,92 @@ export class MayachainQuery {
 
     return {
       count: actionsResume.count ? Number(actionsResume.count) : 0,
-      swaps: actionsResume.actions.map((action) => {
-        let transaction: Transaction | undefined = action.out.filter((out) => out.txID !== '')[0] // For non to protocol asset swap
-        if (!transaction) {
-          // For to protocol asset swap
-          transaction = action.out.sort((out1, out2) => Number(out2.coins[0].amount) - Number(out1.coins[0].amount))[0]
-        }
-        return {
-          date: new Date(Number(action.date) / 10 ** 6),
-          status: action.status,
-          inboundTx: {
+      swaps: actionsResume.actions
+        // Merge duplicated swaps in just one
+        .reduce((prev, current) => {
+          const index = prev.findIndex((action) => action.in[0].txID === current.in[0].txID)
+          if (index === -1) return [...prev, current]
+
+          for (let i = 0; i < current.in.length; i++) {
+            prev[index].in[i].coins[0].amount = String(
+              Number(prev[index].in[i].coins[0].amount) + Number(current.in[i].coins[0].amount),
+            )
+          }
+          return prev
+        }, [] as Action[])
+        .map((action) => {
+          const inboundTx: TransactionAction = {
             hash: action.in[0].txID,
             address: action.in[0].address,
             amount: getCryptoAmount(assetDecimals, action.in[0].coins[0].asset, action.in[0].coins[0].amount),
-          },
-          outboundTx: transaction
-            ? {
-                hash: transaction.txID,
-                address: transaction.address,
-                amount: getCryptoAmount(assetDecimals, transaction.coins[0].asset, transaction.coins[0].amount),
-              }
-            : undefined,
-        }
-      }),
+          }
+
+          const fromAsset: CompatibleAsset = inboundTx.amount.asset
+          const toAsset: CompatibleAsset = this.getAssetFromMemo(
+            (action.metadata.swap as SwapMetadata).memo,
+            poolDetails,
+          )
+
+          if (action.status === 'pending') {
+            return {
+              date: new Date(Number(action.date) / 10 ** 6),
+              status: 'pending',
+              fromAsset,
+              toAsset,
+              inboundTx,
+            }
+          }
+
+          const transaction: Transaction = action.out
+            .filter((out) => out.coins[0].asset === assetToString(toAsset))
+            .sort((out1, out2) => Number(out2.coins[0].amount) - Number(out1.coins[0].amount))[0]
+
+          return {
+            date: new Date(Number(action.date) / 10 ** 6),
+            status: 'success',
+            fromAsset,
+            toAsset,
+            inboundTx,
+            outboundTx: {
+              hash: transaction.txID,
+              address: transaction.address,
+              amount: getCryptoAmount(assetDecimals, transaction.coins[0].asset, transaction.coins[0].amount),
+            },
+          }
+        }),
     }
   }
 
+  private getAssetFromMemo(memo: string, pools: PoolDetail[]): CompatibleAsset {
+    const getAssetFromAliasIfNeeded = (alias: string, pools: PoolDetail[]): string => {
+      let delimiter: string = TOKEN_ASSET_DELIMITER
+
+      if (alias.includes(TRADE_ASSET_DELIMITER)) {
+        delimiter = TRADE_ASSET_DELIMITER
+      } else if (alias.includes(SYNTH_ASSET_DELIMITER)) {
+        delimiter = SYNTH_ASSET_DELIMITER
+      }
+
+      const splitedAlias = alias.split(delimiter)
+      const pool = pools.find((pool) => pool.asset.includes(`${splitedAlias[0]}.${splitedAlias[1].split('-')[0]}`))
+
+      if (pool) return pool.asset.replace('.', delimiter)
+
+      return alias
+    }
+
+    const attributes = memo.split(':')
+    if (!attributes[0]) throw Error(`Invalid memo: ${memo}`)
+
+    switch (attributes[0]) {
+      case 'SWAP':
+      case '=':
+        if (!attributes[1]) throw Error('Asset not defined')
+        return assetFromStringEx(getAssetFromAliasIfNeeded(attributes[1], pools)) as CompatibleAsset
+      default:
+        throw Error(`Get asset from memo unsupported for ${attributes[0]} operation`)
+    }
+  }
   /**
    * Get the MAYANames owned by an address
    * @param {Address} owner - Thorchain address