+++ packages/xchain-thorchain-query/src/thorchain-query.ts
@@ -1,4 +1,4 @@
-import { Transaction } from '@xchainjs/xchain-midgard'
+import { SwapMetadata, Transaction } from '@xchainjs/xchain-midgard'
 import { LastBlock, TradeAccountResponse, TradeUnitResponse } from '@xchainjs/xchain-thornode'
 import {
   Address,
@@ -6,7 +6,10 @@ import {
   AssetCryptoAmount,
   Chain,
   CryptoAmount,
+  SYNTH_ASSET_DELIMITER,
   SynthAsset,
+  TOKEN_ASSET_DELIMITER,
+  TRADE_ASSET_DELIMITER,
   TokenAsset,
   TradeAsset,
   assetAmount,
@@ -62,6 +65,7 @@ import {
   TradeAssetUnits,
   TradeAssetUnitsParams,
   TradeAssetsUnitsParams,
+  TransactionAction,
   TxDetails,
   UnitData,
   WithdrawLiquidityPosition,
@@ -1401,27 +1405,40 @@ export class ThorchainQuery {
     return {
       count: actionsResume.count ? Number(actionsResume.count) : 0,
       swaps: actionsResume.actions.map((action) => {
-        let transaction: Transaction | undefined = action.out.filter((out) => out.txID !== '')[0] // For non to protocol asset swap
-
-        if (!transaction) {
-          // For to protocol asset swap
-          transaction = action.out.sort((out1, out2) => Number(out2.coins[0].amount) - Number(out1.coins[0].amount))[0]
+        const inboundTx: TransactionAction = {
+          hash: action.in[0].txID,
+          address: action.in[0].address,
+          amount: getInboundCryptoAmount(pools, action.in[0].coins[0].asset, action.in[0].coins[0].amount),
         }
+
+        const fromAsset: CompatibleAsset = inboundTx.amount.asset
+        const toAsset: CompatibleAsset = this.getAssetFromMemo((action.metadata.swap as SwapMetadata).memo, pools)
+
+        if (action.status === 'pending') {
+          return {
+            date: new Date(Number(action.date) / 10 ** 6),
+            status: 'pending',
+            fromAsset,
+            toAsset,
+            inboundTx,
+          }
+        }
+
+        const transaction: Transaction = action.out
+          .filter((out) => out.coins[0].asset === assetToString(toAsset))
+          .sort((out1, out2) => Number(out2.coins[0].amount) - Number(out1.coins[0].amount))[0]
+
         return {
           date: new Date(Number(action.date) / 10 ** 6),
-          status: action.status,
-          inboundTx: {
-            hash: action.in[0].txID,
-            address: action.in[0].address,
-            amount: getInboundCryptoAmount(pools, action.in[0].coins[0].asset, action.in[0].coins[0].amount),
+          status: 'success',
+          fromAsset,
+          toAsset,
+          inboundTx,
+          outboundTx: {
+            hash: transaction.txID,
+            address: transaction.address,
+            amount: getInboundCryptoAmount(pools, transaction.coins[0].asset, transaction.coins[0].amount),
           },
-          outboundTx: transaction
-            ? {
-                hash: transaction.txID,
-                address: transaction.address,
-                amount: getInboundCryptoAmount(pools, transaction.coins[0].asset, transaction.coins[0].amount),
-              }
-            : undefined,
         }
       }),
     }
@@ -1587,5 +1604,52 @@ export class ThorchainQuery {
         lastWithdrawHeight: position.last_withdraw_height,
       }
     })
+  }
+
+  private getAssetFromMemo(memo: string, pools: Record<string, LiquidityPool>): CompatibleAsset {
+    const getAssetFromAliasIfNeeded = (alias: string, pools: Record<string, LiquidityPool>): string => {
+      const nativeAlias = new Map<string, string>([
+        ['r', 'THOR.RUNE'],
+        ['rune', 'THOR.RUNE'],
+        ['b', 'BTC.BTC'],
+        ['e', 'ETH.ETH'],
+        ['g', 'GAIA.ATOM'],
+        ['d', 'DOGE.DOGE'],
+        ['l', 'LTC.LTC'],
+        ['c', 'BCH.BCH'],
+        ['a', 'AVAX.AVAX'],
+        ['s', 'BSC.BNB'],
+      ])
+
+      const nativeAsset = nativeAlias.get(alias.toLowerCase())
+      if (nativeAsset) return nativeAsset
+
+      let delimiter: string = TOKEN_ASSET_DELIMITER
+
+      if (alias.includes(TRADE_ASSET_DELIMITER)) {
+        delimiter = TRADE_ASSET_DELIMITER
+      } else if (alias.includes(SYNTH_ASSET_DELIMITER)) {
+        delimiter = SYNTH_ASSET_DELIMITER
+      }
+
+      const splitedAlias = alias.split(delimiter)
+      const poolId = Object.keys(pools).find((pool) => pool === `${splitedAlias[0]}.${splitedAlias[1].split('-')[0]}`)
+
+      if (poolId) return pools[poolId].assetString.replace('.', delimiter)
+
+      return alias
+    }
+
+    const attributes = memo.split(':')
+    if (!attributes[0]) throw Error(`Invalid memo: ${memo}`)
+
+    switch (attributes[0]) {
+      case 'SWAP':
+      case '=':
+        if (!attributes[1]) throw Error('Asset not defined')
+        return assetFromStringEx(getAssetFromAliasIfNeeded(attributes[1], pools)) as CompatibleAsset
+      default:
+        throw Error(`Get asset from memo unsupported for ${attributes[0]} operation`)
+    }
   }
 }