+++ packages/xchain-thorchain-query/src/types.ts
@@ -488,15 +488,27 @@ export type TransactionAction = {
   amount: CryptoAmount
 }
 
+export type SuccessSwap = {
+  date: Date
+  fromAsset: CompatibleAsset
+  toAsset: CompatibleAsset
+  status: 'success'
+  inboundTx: TransactionAction
+  outboundTx: TransactionAction
+}
+
+export type PendingSwap = {
+  date: Date
+  fromAsset: CompatibleAsset
+  toAsset: CompatibleAsset
+  status: 'pending'
+  inboundTx: TransactionAction
+}
+
 /**
  * Swap resume
  */
-export type Swap = {
-  date: Date
-  status: 'success' | 'pending'
-  inboundTx: TransactionAction
-  outboundTx?: TransactionAction
-}
+export type Swap = SuccessSwap | PendingSwap
 
 /**
  * Swap history