+++ packages/xchain-base/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## 0.0.6
+
+### Patch Changes
+
+- 33bfa40: Rollup update to latest version.
+- Updated dependencies [33bfa40]
+  - @xchainjs/xchain-evm-providers@1.0.8
+  - @xchainjs/xchain-client@1.0.6
+  - @xchainjs/xchain-util@1.0.5
+  - @xchainjs/xchain-evm@1.0.10
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [73b68ed]
+  - @xchainjs/xchain-util@1.0.4
+  - @xchainjs/xchain-client@1.0.5
+  - @xchainjs/xchain-evm@1.0.9
+  - @xchainjs/xchain-evm-providers@1.0.7
+
 ## 0.0.4
 
 ### Patch Changes