+++ packages/xchain-mayamidgard-query/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 0.1.22
+
+### Patch Changes
+
+- 33bfa40: Rollup update to latest version.
+- Updated dependencies [33bfa40]
+  - @xchainjs/xchain-mayamidgard@0.1.7
+  - @xchainjs/xchain-client@1.0.6
+  - @xchainjs/xchain-util@1.0.5
+
+## 0.1.21
+
+### Patch Changes
+
+- Updated dependencies [73b68ed]
+  - @xchainjs/xchain-util@1.0.4
+  - @xchainjs/xchain-client@1.0.5
+
 ## 0.1.20
 
 ### Patch Changes